Add files via upload
The code performs a numerical analysis of temperature evolution T(t), described by the equation 
𝑑𝑇/𝑑𝑡=−0.0059𝑇, and evaluates errors in numerical methods. The user can choose between visualizing errors (local and Runge) or solving the equation using the second-order Adams-Bashforth method.

In the graph mode (g/G), it plots the comparison of local errors and Runge errors for data with a time step Δ𝑇=100. In calculation mode (c/C), the differential equation is solved numerically: the first step uses Euler's method, and subsequent steps use the Adams-Bashforth method. Results are visualized as a temperature evolution graph over time.